Argument call spreading
There are 4 kinds of bytecode calls in Teascript VM: `CALL`, `INVOKE`, `SUPER` and `NEW`.

Spread of function arguments turns out to be kind of tricky. Currently, the implementation is basically self-modifying bytecode.

Is it bad? Not sure, but it does work! Surely there's another way to implement it, but for now I think it's good enough?

Well, there are also other parts to fix, like optional function arguments. And later on converting the VM to register based.

Let's see how it goes...

+++ src/tea_obj.h
@@ -162,7 +162,7 @@ typedef struct
     uint8_t numparams;  /* Number of parameters */
     uint8_t numopts; /* Number of optional parameters */
     uint8_t sizeuv;  /* Number of upvalues */
-    uint8_t max_slots;  /* Max stack size used by the function */
+    uint16_t max_slots;  /* Max stack size used by the function */
     uint8_t flags;   /* Miscellaneous flags */
     uint32_t sizept;    /* Total size including colocated arrays */
     uint32_t sizebc;  /* Number of bytecode instructions */
